Split nodes not descendant from AlogicAST to Misc.scala
