+++ src/main/scala/alogic/ast/AlogicAST.scala
@@ -102,21 +102,6 @@ case class GotoStmt(target: String) extends AlogicStmt // tODO: take DottedName 
 case class StateBlock(state: Int, contents: List[AlogicStmt]) extends AlogicStmt
 case class GotoState(state: Int) extends AlogicStmt
 
-// Case classes do not support inheritance
-// Use sealed trait to force compiler to detect all cases are covered
-// Use different abstract classes to decompose problem.
-
-//
-// // Declaration used for top-level and function declarations
-sealed trait Declaration
-// TODO: Add separate Decl for Arrays, it just makes a lot of things simpler
-case class VarDeclaration(decltype: AlogicType, id: AlogicAST, init: Option[AlogicExpr]) extends Declaration
-case class ParamDeclaration(decltype: AlogicType, id: String, init: AlogicExpr) extends Declaration
-case class ConstDeclaration(decltype: AlogicType, id: String, init: AlogicExpr) extends Declaration
-case class VerilogDeclaration(decltype: AlogicType, id: AlogicAST) extends Declaration
-case class OutDeclaration(synctype: SyncType, decltype: AlogicType, name: String) extends Declaration
-case class InDeclaration(synctype: SyncType, decltype: AlogicType, name: String) extends Declaration
-
 // // AlogicAST used for abstract syntax nodes
 case class Connect(start: AlogicAST, end: List[AlogicAST]) extends AlogicAST
 case class Instantiate(id: String, module: String, args: List[AlogicAST]) extends AlogicAST
@@ -125,20 +110,3 @@ case class CombinatorialCaseLabel(cond: List[AlogicExpr], body: AlogicStmt) exte
 
 // Types removed by MakeStates
 case class ControlCaseLabel(cond: List[AlogicExpr], body: AlogicStmt) extends AlogicAST
-
-// AlogicType used to define the allowed types
-sealed trait AlogicType
-case class IntType(signed: Boolean, size: Int) extends AlogicType
-case class IntVType(signed: Boolean, args: List[AlogicExpr]) extends AlogicType // variable number of bits definition
-case class Struct(fields: Map[String, AlogicType]) extends AlogicType
-case object State extends AlogicType // Type with enough bits to hold state variable
-
-// SyncType for allowed port types
-sealed trait SyncType
-case object SyncReadyBubble extends SyncType
-case object SyncReady extends SyncType
-case object SyncAccept extends SyncType
-case object Sync extends SyncType
-case object WireSyncAccept extends SyncType
-case object WireSync extends SyncType
-case object Wire extends SyncType
